refactor: 지원현황 페이지 및 wrapper 리펙토링
