+++ src/pages/wrappers/EditStudyWrapper.tsx
@@ -1,16 +1,6 @@
-import { useNavigate, useParams } from "react-router-dom";
 import EditStudyPage from "../EditStudyPage";
+import StudyPageWrapper from "../../components/wrappers/StudyPageWrapper";
 
-function EditStudyWrapper() {
-    const { studyId } = useParams<{ studyId: string }>();
-    const navigate = useNavigate();
-
-    return (
-        <EditStudyPage
-            studyId={Number(studyId)}
-            onBack={() => navigate(`/detail/${studyId}`)}
-        />
-    );
+export default function EditStudyWrapper() {
+  return <StudyPageWrapper PageComponent={EditStudyPage} />;
 }
-
-export default EditStudyWrapper;
